homework_1/project/requirements.txt
torch
torchvision
lightning
torchmetrics
wandb
optuna
python-dotenv
